Prints all combinations of 2 2-digits numbers
The numbers are ranged from 0 to 99
The two numbers are separated by a space
All numbers are printed with two digits. 1 is printed as 01
The combinations of numbers are separated by comma, followed by a space
The combinations of numbers are printed in ascending order.
We only used putchar function. And this function is used 8 times max.
00 01 and 01 00 are considered as the same combination of the numbers 0
and 1