Fix: Compile script 4-isalpha.c successfully
This fixes the broken behavior of the compilation process.

BREAKING CHANGES
Before this fix, this script used an external function; _islower()
The checker was configure to compile with inputs scripts that do not
include 3-islower.c
Behavior changes from including 3-islower.c in the compilation process
to not including this script anymore. We just copy the function in
_islower(inside 3-islower.c) into our target script.